AKHIRNYA KELAR JUGA KITA

+++ database/migrations/2022_12_13_070752_create_products_table.php
@@ -18,7 +18,7 @@ class CreateProductsTable extends Migration
             // $table->timestamps();
             $table->string('name');
             $table->string('size');
-            $table->float('price');
+            $table->unsignedBigInteger('price');
             $table->integer('qty');
             $table->string('image');
         });
